pyproject.toml
[project]
name = "anaya"
version = "1.0.0"
description = "Compliance-as-code engine — GitHub App that scans PRs against policy rule packs"
readme = "README.md"
license = { text = "AGPL-3.0" }
requires-python = ">=3.12"
authors = [{ name = "Anaya Team" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "PyJWT[crypto]>=2.9",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "PyYAML>=6.0",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "typer>=0.14",
    "rich>=13.9",
    "tree-sitter>=0.24",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "jsonschema>=4.23",
]

[project.scripts]
anaya = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["anaya", "cli"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
